Saving progress on new Jane bot.
Need to make dir selection in placeShip routine prefer trying directions
that overlap hits first.

+++ src/bots/Hal9000.h
@@ -19,8 +19,8 @@ public: // constructors
 
 //-----------------------------------------------------------------------------
 protected: // Bot implementation
-  void frenzyScore(const Board&, ScoredCoordinate&, const unsigned) override;
-  void searchScore(const Board&, ScoredCoordinate&, const unsigned) override;
+  void frenzyScore(const Board&, ScoredCoordinate&, const double) override;
+  void searchScore(const Board&, ScoredCoordinate&, const double) override;
 };
 
 } // namespace xbs
